fix external database startup failed.
